requirements.txt
--index-url https://pypi.org/simple/
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
decorator==4.4.2
distro==1.9.0
fastapi>=0.104.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.3
idna==3.10
ImageHash==4.3.2
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
llvmlite==0.44.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai==1.98.0
openai-whisper==20250625
opensearch-py==2.7.1
opencv-python==4.12.0.88
orjson==3.11.1
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
proglog==0.1.12
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.17
PyWavelets==1.8.0
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scenedetect==0.6.6
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.0.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette>=0.40.0,<0.42.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.54.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.34.0
yarl==1.20.1
zstandard==0.23.0
